pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "kws-gscv2"
version = "1.0.0"
description = "Keyword Spotting on Google Speech Commands v2 (12-class)"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "PyYAML>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.scripts]
kws-train = "train:main"
kws-prepare-data = "scripts.prepare_data:main"
